% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateObject.R
\name{CreateObject}
\alias{CreateObject}
\title{Create a Seurat object with transcript and gene assays plus transcript-level QC metrics}
\usage{
CreateObject(
  tmat,
  gmat,
  gtf.df,
  project = "SeuratProject",
  transcript.assay = "transcript",
  gene.assay = "gene",
  protein.coding.label = "protein_coding",
  mt.pattern = "^MT-|^mt-",
  verbose = FALSE
)
}
\arguments{
\item{tmat}{A sparse matrix with transcripts/isoforms as rows and cells as columns.}

\item{gmat}{A sparse matrix with genes as rows and cells as columns.}

\item{gtf.df}{A data.frame aligned to \code{tmat} with at least columns \code{transcript_id} and a gene label column.}

\item{project}{Character(1). Seurat project name (default = "SeuratProject").}

\item{transcript.assay}{Character(1). Name for transcript assay (default = "transcript")}

\item{gene.assay}{Character(1). Name for gene assay (default = "gene")}

\item{protein.coding.label}{Character(1). Label indicating protein-coding transcripts in \code{gtf.df$transcript.type} (default = "protein_coding").}

\item{mt.pattern}{Character(1). Regex used to identify mitochondrial genes in the gene assay. Default matches common human and mouse conventions (default = '^MT-|^mt-')"}

\item{verbose}{Logical(1). If TRUE, print progress messages (default = FALSE).}
}
\value{
A Seurat object with transcript and gene assays plus added QC metadata.
}
\description{
Create a Seurat object with transcript and gene assays plus transcript-level QC metrics
}
