% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_psi.R
\name{.parse_ioe_files}
\alias{.parse_ioe_files}
\title{Parse one or more SUPPA2 IOE files into an event table}
\usage{
.parse_ioe_files(ioe_files)
}
\arguments{
\item{ioe_files}{Character vector of file paths.}
}
\value{
A data.frame with columns: \code{event_id}, \code{gene_id},
\code{chr}, \code{strand}, \code{event_type},
\code{inclusion_transcripts}, \code{exclusion_transcripts}.
}
\description{
Parse one or more SUPPA2 IOE files into an event table
}
\keyword{internal}
