% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterBarcodes.R
\name{FilterBarcodes}
\alias{FilterBarcodes}
\title{Filter a transcript matrix by top-ranked barcodes by retaining the top \code{cells} barcodes ranked by total counts per barcode.}
\usage{
FilterBarcodes(tmat, cells)
}
\arguments{
\item{tmat}{A sparse matrix with transcripts/isoforms as rows and cells as columns.}

\item{cells}{Integer(1). Number of top barcodes with positive counts to retain.}
}
\value{
A filtered sparse matrix containing only the retained barcodes.
}
\description{
Filter a transcript matrix by top-ranked barcodes by retaining the top \code{cells} barcodes ranked by total counts per barcode.
}
