% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{GetEventData}
\alias{GetEventData}
\alias{GetEventData,MatisseObject-method}
\title{Get splice event annotation table}
\usage{
GetEventData(object, ...)

\S4method{GetEventData}{MatisseObject}(object, ...)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{data.frame} of splice event metadata.
}
\description{
Get splice event annotation table
}
