% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{GetInclusionCounts}
\alias{GetInclusionCounts}
\alias{GetInclusionCounts,MatisseObject-method}
\title{Get inclusion read count matrix}
\usage{
GetInclusionCounts(object, ...)

\S4method{GetInclusionCounts}{MatisseObject}(object, ...)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse matrix (cells x events) of inclusion read counts.
}
\description{
Retrieves inclusion counts from the \code{"counts"} layer of the
\code{"psi"} \code{Assay5}.
}
