% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{GetJunctionData}
\alias{GetJunctionData}
\alias{GetJunctionData,MatisseObject-method}
\title{Get junction annotation table}
\usage{
GetJunctionData(object, ...)

\S4method{GetJunctionData}{MatisseObject}(object, ...)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{data.frame} of junction metadata.
}
\description{
Get junction annotation table
}
