% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{GetSeurat}
\alias{GetSeurat}
\alias{GetSeurat,MatisseObject-method}
\title{Get the embedded Seurat object}
\usage{
GetSeurat(object, ...)

\S4method{GetSeurat}{MatisseObject}(object, ...)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{Seurat} object.
}
\description{
Get the embedded Seurat object
}
