% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTranscriptMatrix.R
\name{LoadTranscriptMatrix}
\alias{LoadTranscriptMatrix}
\title{Load single-cell transcript-level RNA-seq data as a sparse matrix.}
\usage{
LoadTranscriptMatrix(
  barcode.file,
  feature.file,
  matrix.file,
  cell.column = 1L,
  feats.column = 1L,
  transpose = FALSE,
  make.unique.features = FALSE
)
}
\arguments{
\item{barcode.file}{Character(1). Path to barcode file.}

\item{feature.file}{Character(1). Path to feature file.}

\item{matrix.file}{Character(1). Path to Matrix Market file.}

\item{cell.column}{Integer(1). Column of \code{barcode.file} to use for cell names (default = 1).}

\item{feats.column}{Integer(1). Column of \code{feature.file} to use for feature names (default = 1).}

\item{transpose}{Logical(1). If TRUE, transpose the matrix after reading (default = FALSE).}

\item{make.unique.features}{Logical(1). If TRUE, make duplicated feature names unique (default = FALSE).}
}
\value{
A \code{Matrix::dgCMatrix} with features as rows and cells as columns.
}
\description{
Load single-cell transcript-level RNA-seq data as a sparse matrix.
}
