% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{GetJunctionCounts}
\alias{GetJunctionCounts}
\alias{GetJunctionCounts,MatisseObject-method}
\title{Get raw junction count matrix}
\usage{
GetJunctionCounts(object, ...)

\S4method{GetJunctionCounts}{MatisseObject}(object, ...)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{...}{Additional arguments (unused).}
}
\value{
A sparse matrix (cells × junctions) of read counts.
}
\description{
Get raw junction count matrix
}
