% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/MatisseObject-methods.R
\name{SetPSI}
\alias{SetPSI}
\alias{SetPSI,MatisseObject-method}
\title{Set the PSI matrix}
\usage{
SetPSI(object, value)

\S4method{SetPSI}{MatisseObject}(object, value)
}
\arguments{
\item{object}{A \code{MatisseObject}.}

\item{value}{A sparse matrix (cells × events) of PSI values.}
}
\value{
The updated \code{MatisseObject}.
}
\description{
Replaces the \code{"data"} layer of the \code{"psi"} \code{Assay5}
inside the embedded Seurat object.
}
